[#948] add resource_modified() to migration file

+++ hs_geo_raster_resource/migrations/custom_migration_for_tif_to_vrt_20160223.py
@@ -9,6 +9,7 @@ from django.core.files.uploadedfile import UploadedFile
 
 
 from hs_core import hydroshare
+from hs_core.hydroshare.utils import resource_modified
 from hs_geo_raster_resource.models import RasterResource
 from hs_geo_raster_resource.receivers import create_vrt_file
 
@@ -18,20 +19,20 @@ def migrate_tif_file(apps, schema_editor):
     log = logging.getLogger()
     for res in RasterResource.objects.all():
 
-        res_file = res.files.all().first()
-        vrt_file_path, temp_dir = create_vrt_file(res_file.resource_file)
-        if os.path.isfile(vrt_file_path):
-            files = (UploadedFile(file=open(vrt_file_path, 'r'), name=os.path.basename(vrt_file_path)))
-            hydroshare.add_resource_files(res.short_id, files)
-            print 'Tif file conversion to VRT successful for resource:ID:{} Title:{}'.format(res.short_id, res.metadata.title.value)
-            log.info('Tif file conversion to VRT successful for resource:ID:{} '
-                     'Title:{}'.format(res.short_id, res.metadata.title.value))
-        else:
-            print 'Tif file conversion to VRT unsuccessful for resource:ID:{} Title:{}'.format(res.short_id, res.metadata.title.value)
-            log.error('Tif file conversion to VRT unsuccessful for resource:ID:{} '
-                      'Title:{}'.format(res.short_id, res.metadata.title.value))
-        if os.path.isdir(temp_dir):
-            shutil.rmtree(temp_dir)
+        if len(res.files.all()) == 1:
+            res_file = res.files.all().first()
+            vrt_file_path, temp_dir = create_vrt_file(res_file.resource_file)
+            if os.path.isfile(vrt_file_path):
+                files = (UploadedFile(file=open(vrt_file_path, 'r'), name=os.path.basename(vrt_file_path)))
+                hydroshare.add_resource_files(res.short_id, files)
+                resource_modified(res, res.creator)
+                log.info('Tif file conversion to VRT successful for resource:ID:{} '
+                         'Title:{}'.format(res.short_id, res.metadata.title.value))
+            else:
+                log.error('Tif file conversion to VRT unsuccessful for resource:ID:{} '
+                          'Title:{}'.format(res.short_id, res.metadata.title.value))
+            if os.path.isdir(temp_dir):
+                shutil.rmtree(temp_dir)
 
 
 def undo_migrate_tif_file(apps, schema_editor):
